+++ system/partners.php
@@ -1,0 +1,5 @@
+<?php
+return [
+  'enabled' => true,
+  'levels' => ['community','certified','strategic'],
+];